reuse db connection handling and improve it
