+++ src/server/db/index.js
@@ -3,6 +3,9 @@
 var _ = require('lodash');
 var P = require('../../util/promise');
 var mongoose = require('mongoose');
+var mongoose_logger = require('../../util/mongoose_logger');
+var dbg = require('../../util/debug_module')(__filename);
+
 
 var LRUCache = require('../../util/lru_cache');
 var Account = require('./account');
@@ -21,6 +24,41 @@ var Pool = require('./pool');
 var TieringPolicy = require('./tiering_policy');
 var dbutils = require('./dbutils');
 // var dbg = require('../util/debug_module')(__filename);
+var debug_mode = (process.env.DEBUG_MODE === 'true');
+var mongoose_connected = false;
+var mongoose_timeout = null;
+
+// connect to the database
+if (debug_mode) {
+    mongoose.set('debug', mongoose_logger(dbg.log0.bind(dbg)));
+}
+
+mongoose.connection.on('connected', function() {
+    // call ensureIndexes explicitly for each model
+    console.log('mongoose connection connected');
+    mongoose_connected = true;
+    return P.all(_.map(mongoose.modelNames(), function(model_name) {
+        return P.npost(mongoose.model(model_name), 'ensureIndexes');
+    }));
+});
+
+mongoose.connection.on('error', function(err) {
+    mongoose_connected = false;
+    console.error('mongoose connection error:', err);
+    if (!mongoose_timeout) {
+        mongoose_timeout = setTimeout(mongoose_conenct, 5000);
+    }
+
+});
+
+mongoose.connection.on('disconnected', function () {
+    mongoose_connected = false;
+    console.error('mongoose connection disconnected');
+    if (!mongoose_timeout) {
+        mongoose_timeout = setTimeout(mongoose_conenct, 5000);
+    }
+});
+
 
 /**
  *
@@ -54,6 +92,7 @@ module.exports = {
 
     check_not_found: check_not_found,
     check_not_deleted: check_not_deleted,
+    mongoose_conenct: mongoose_conenct,
     check_already_exists: check_already_exists,
     is_err_exists: is_err_exists,
     obj_ids_difference: obj_ids_difference,
@@ -165,6 +204,20 @@ module.exports = {
     }),
 };
 
+
+
+function mongoose_conenct() {
+    console.error('mongoose attempt to connect');
+
+    clearTimeout(mongoose_timeout);
+    mongoose_timeout = null;
+    if (!mongoose_connected) {
+        mongoose.connect(
+            process.env.MONGOHQ_URL ||
+            process.env.MONGOLAB_URI ||
+            'mongodb://localhost/nbcore');
+    }
+}
 
 function check_not_found(req, entity) {
     return function(doc) {
