+++ src/server/web_server.js
@@ -17,7 +17,6 @@ var P = require('../util/promise');
 var path = require('path');
 var http = require('http');
 var https = require('https');
-var mongoose = require('mongoose');
 var express = require('express');
 var express_favicon = require('serve-favicon');
 var express_morgan_logger = require('morgan');
@@ -29,14 +28,13 @@ var express_compress = require('compression');
 var util = require('util');
 var config = require('../../config.js');
 var dbg = require('../util/debug_module')(__filename);
-var mongoose_logger = require('../util/mongoose_logger');
 var time_utils = require('../util/time_utils');
 var pem = require('../util/pem');
 var multer = require('multer');
 var fs = require('fs');
 var cluster = require('cluster');
 var pkg = require('../../package.json');
-
+var db = require('../server/db');
 
 
 
@@ -62,44 +60,9 @@ process.env.ADDRESS = process.env.ADDRESS || 'http://localhost:5001';
 
 var rootdir = path.join(__dirname, '..', '..');
 var dev_mode = (process.env.DEV_MODE === 'true');
-var debug_mode = (process.env.DEBUG_MODE === 'true');
-var mongoose_connected = false;
-var mongoose_timeout = null;
 
-// connect to the database
-if (debug_mode) {
-    mongoose.set('debug', mongoose_logger(dbg.log0.bind(dbg)));
-}
 
-mongoose.connection.once('open', function() {
-    // call ensureIndexes explicitly for each model
-    mongoose_connected = true;
-    return P.all(_.map(mongoose.modelNames(), function(model_name) {
-        return P.npost(mongoose.model(model_name), 'ensureIndexes');
-    }));
-});
-
-mongoose.connection.on('error', function(err) {
-    mongoose_connected = false;
-    console.error('mongoose connection error:', err);
-    if (!mongoose_timeout) {
-        mongoose_timeout = setTimeout(mongoose_conenct, 5000);
-    }
-
-});
-
-function mongoose_conenct() {
-    clearTimeout(mongoose_timeout);
-    mongoose_timeout = null;
-    if (!mongoose_connected) {
-        mongoose.connect(
-            process.env.MONGOHQ_URL ||
-            process.env.MONGOLAB_URI ||
-            'mongodb://localhost/nbcore');
-    }
-}
-
-mongoose_conenct();
+db.mongoose_conenct();
 
 // create express app
 var app = express();
